Updated delete.php for authors.

+++ api/authors/delete.php
@@ -1,29 +1,40 @@
 <?php
-    // Include headers
-    require_once '../../config/headers.php';
+  // Include headers
+  require_once '../../config/headers.php';
 
-    include_once '../../config/Database.php';
-    include_once '../../models/Author.php';
-    // Instantiate DB & connect
-    $database = new Database();
-    $db = $database->connect();
+  // Include necessary files
+  require_once '../../config/Database.php';
+  require_once '../../models/Author.php';
 
-    // Instantiate blog post object
-    $author = new Author($db);
+  // Instantiate DB & connect
+  $database = new Database();
+  $db = $database->connect();
 
-    // Get raw posted data
-    $data = json_decode(file_get_contents("php://input"));
+  // Instantiate Author model
+  $author = new Author($db);
 
-    // Set ID to UPDATE
-    $author->id = $data->id;
+  // Get raw posted data
+  $data = json_decode(file_get_contents("php://input"));
 
-    // Delete post
-    if($author->delete()) 
-    {
-      echo json_encode((["message" => "Author deleted"]));
-    } 
+  // Validate input
+  if (!isset($data->id) || empty($data->id)) 
+  {
+      echo json_encode(["message" => "Author not deleted"]);
+      exit();
+  }
 
-    else 
-    {
-      echo json_encode((["message" => "Author not deleted"]));
-    }
+  // Set ID for deletion
+  $author->id = (int) $data->id;
+
+  // Attempt to delete
+  if ($author->delete()) 
+  {
+      echo json_encode(["message" => "Author deleted"]);
+  } 
+
+  else 
+  {
+      echo json_encode(["message" => "Author not deleted"]);
+  }
+
+  exit();
